chore(refactor): Convert class component to functional
