+++ src/Components/PresentationalComponents/SelectAllCheckbox.js
@@ -1,0 +1,102 @@
+import { Checkbox, Dropdown, DropdownItem, DropdownToggle } from '@patternfly/react-core';
+import propTypes from 'prop-types';
+import React, { useState } from 'react';
+
+const SelectAllCheckbox = (props) => {
+    const [isOpen, setIsOpen] = useState(false);
+
+    const { selectedItems, items, multiRow } = props;
+    const { meta, data } = items;
+    const checkboxlabel = Boolean(selectedItems) && `${selectedItems} selected`;
+    const selectPageNumber = multiRow ? data.length / 2 : data.length;
+    const itemLabel = number => (number > 1 && 'items') || 'item';
+
+    const handleOnToggle = isOpen => {
+        setIsOpen(isOpen);
+    };
+
+    const handleOnSelect = () => {
+        setIsOpen(!isOpen);
+    };
+
+    const handleSelectPage = () => {
+        const { items: { data },  selectorHandler } = props;
+        const items = data.filter(item => item.id).map(item => item.id);
+        selectorHandler(true, items);
+    };
+
+    const handleUnselectAll = () => {
+        const { selectorHandler } = props;
+        selectorHandler(false, undefined);
+    };
+
+    const handleSelectAll = () => {
+        const { items: { meta }, selectorHandler, fetchResource } = props;
+        // eslint-disable-next-line camelcase
+        let { payload } = fetchResource && fetchResource({ page_size: meta.total_items, page: 1 });
+        payload &&
+            payload.then(({ data: response }) => {
+                const items = response.map(item => item.id);
+                selectorHandler(true, items);
+            });
+    };
+
+    const handleOnCheckboxChange = () => {
+        const { selectedItems } = props;
+        if (selectedItems === 0) {
+            handleSelectAll();
+        } else {
+            handleUnselectAll();
+        }
+    };
+
+    const dropdownItems = [
+        <DropdownItem key="none" onClick={handleUnselectAll}>
+            Select none (0 items)
+        </DropdownItem>,
+        <DropdownItem key="page" onClick={handleSelectPage}>
+            Select page ({selectPageNumber} {itemLabel(selectPageNumber)})
+        </DropdownItem>,
+        <DropdownItem key="all" onClick={handleSelectAll}>
+            Select all ({meta.total_items} {itemLabel(meta.total_items)})
+        </DropdownItem>
+    ];
+
+    return (
+        <Dropdown
+            onSelect={handleOnSelect}
+            toggle={
+                <DropdownToggle
+                    splitButtonItems={[
+                        <Checkbox
+                            key={'selectAllcheckbox'}
+                            isChecked={meta.total_items === selectedItems ? true : selectedItems === 0 ? false : null}
+                            onChange={handleOnCheckboxChange}
+                        />,
+                        checkboxlabel
+                    ]}
+                    onToggle={handleOnToggle}
+                />
+            }
+            isOpen={isOpen}
+            dropdownItems={dropdownItems}
+        />
+    );
+};
+
+SelectAllCheckbox.propTypes = {
+    selectorHandler: propTypes.func,
+    fetchResource: propTypes.func,
+    selectedItems: propTypes.object,
+    items: propTypes.object,
+    multiRow: propTypes.bool
+};
+
+SelectAllCheckbox.defaultProps = {
+    impact: 'Unknown',
+    selectedItems: {},
+    items: {},
+    multiRow: false
+};
+
+export default SelectAllCheckbox;
