+++ src/Components/SmartComponents/CVEs/VulnerabilitiesTableToolbar.js
@@ -22,7 +22,7 @@ import FilterChips from '../../PresentationalComponents/Filters/FilterChips';
 import Filters from '../../PresentationalComponents/Filters/Filters';
 import PaginationWrapper from '../../PresentationalComponents/PaginationWrapper/PaginationWrapper';
 import Remediation from '../Remediation/Remediation';
-import SelectAllCheckbox from '../SelectAllCheckbox/SelectAllCheckbox';
+import SelectAllCheckbox from '../../PresentationalComponents/SelectAllCheckbox';
 import { CVETableContext } from './VulnerabilitiesCves';
 
 class VulnerabilitiesToolbarWithContext extends Component {
