+++ src/Components/SmartComponents/SystemCves/SystemCveTableToolbar.js
@@ -20,7 +20,7 @@ import FilterChips from '../../PresentationalComponents/Filters/FilterChips';
 import Filters from '../../PresentationalComponents/Filters/Filters';
 import PaginationWrapper from '../../PresentationalComponents/PaginationWrapper/PaginationWrapper';
 import Remediation from '../Remediation/Remediation';
-import SelectAllCheckbox from '../SelectAllCheckbox/SelectAllCheckbox';
+import SelectAllCheckbox from '../../PresentationalComponents/SelectAllCheckbox';
 import { CVETableContext } from './SystemCves';
 
 class SystemCveToolbarWithContext extends Component {
