+++ src/Components/SmartComponents/Systems/Systems.js
@@ -25,7 +25,7 @@ import {
     kebabItemExcludeSystemAnalysis
 } from '../../PresentationalComponents/DownloadReportKebab/KebabItems';
 import './Systems.scss';
-import SelectAllCheckbox from '../SelectAllCheckbox/SelectAllCheckbox';
+import SelectAllCheckbox from '../../PresentationalComponents/SelectAllCheckbox';
 
 /*eslint camelcase: 0*/
