+++ src/Components/SmartComponents/SystemsExposedTable/SystemsExposedTable.js
@@ -41,7 +41,7 @@ import {
     kebabItemDownloadJSON,
     kebabItemEditStatus
 } from '../../PresentationalComponents/DownloadReportKebab/KebabItems';
-import SelectAllCheckbox from '../SelectAllCheckbox/SelectAllCheckbox';
+import SelectAllCheckbox from '../../PresentationalComponents/SelectAllCheckbox';
 /*eslint camelcase: 0*/
 
 class SystemsExposedTable extends React.Component {
